Remove the old TrafficSignalEditor, cut over to the new variant for both individual and multiple signals
